requirements.txt
torch>=2.2.0
numpy>=1.26.0
pyyaml>=6.0
cnfgen>=0.9.0
matplotlib>=3.8.0
tqdm>=4.66.0
